♻️ Refactor : post/d 페이지 코드 리팩토링
